% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mormar_toughness_est.R
\name{add_mormar_toughness_est}
\alias{add_mormar_toughness_est}
\title{Compute estimated max soil toughness from Atterberg limits}
\usage{
add_mormar_toughness_est(df)
}
\arguments{
\item{df}{A data frame containing (at least) columns named \code{PI} and \code{LL}}
}
\value{
original data frame with a new column titled \code{mormar_toughness_est}
}
\description{
\loadmathjax Takes a data frame  and
calculates the estimated maximum toughness (in kj m^-3) from an empirical
equation published by \href{https://www.sciencedirect.com/science/article/pii/S0169131718301698}{Moreno-Marato and Alonso-Azcarate, 2018}.
Re-arranging their equation yields: \mjsdeqn{\mathrm{T_{max(\frac{kJ}{m^3})}=\frac{\frac{PI}{LL}-0.3397}{0.0077}}}
}
