% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beautify_anova_table.R
\name{beautify_anova_table}
\alias{beautify_anova_table}
\title{Format the output of a model}
\usage{
beautify_anova_table(x)
}
\arguments{
\item{x}{output of \code{car::Anova()}}
}
\value{
tibble with better column names and formatted cell values
}
\description{
Format the output of a model
}
