% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-cleaning-funs.R
\name{both_clean}
\alias{both_clean}
\title{Remove both empty rows and columns; helper called by df_clean based on user input}
\usage{
both_clean(df)
}
\arguments{
\item{df}{Tibble (or data frame)}
}
\description{
Remove both empty rows and columns; helper called by df_clean based on user input
}
