% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_datasheet_path.R
\name{build_datasheet_path}
\alias{build_datasheet_path}
\title{Constructs a path to write an empty delimited file for collecting data}
\usage{
build_datasheet_path(stem, base = NULL, date = NULL, ext = "csv")
}
\arguments{
\item{stem}{character vector of file stems, without extension}

\item{base}{root of file; i.e. directory where it should be saved. Defaults
to \code{here::here()} if not supplied}

\item{date}{date for collection, defaults to today's date}

\item{ext}{file extension, defaults to "csv" but also accepts "tsv" or
"xlsx"; period optional; accepts a character vector of multiple formats if
desired}
}
\value{
character vector of paths with length matching that of the \code{stem}
argument
}
\description{
Used as a helper by my other packages; automatically appends the date to the
end of the file path (prefixed by an underscore)
}
\examples{
# sensible defaults allow for minimal typing
build_datasheet_path(stem = "my-precious-data")

# or write to tab-delimited file
build_datasheet_path(stem = "my-precious-data", ext = "tsv")

# or make many sheets at once using multiple stems and multiple formats

build_datasheet_path(stem = paste("my-precious-data", 1:3, sep = "-"))

# multiple formats can be created with one call
build_datasheet_path(stem = "my-precious-data", ext = c("csv", "tsv"))


}
