% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_drive_url.R
\name{build_drive_url}
\alias{build_drive_url}
\title{Construct a URL Path to Content on Google Drive}
\usage{
build_drive_url(id)
}
\arguments{
\item{id}{Character. Unique Google Drive ID, easily obtained with \code{\link[googledrive:as_id]{googledrive::as_id()}} from the \href{https://googledrive.tidyverse.org}{googledrive} package.}
}
\value{
Character vector with boilerplate url and ID
}
\description{
Makes it simple to access content hosted on Google Drive in output from R Markdown and friends
}
\examples{
\dontrun{
\# save url as a variable
baseball <- build_drive_url("1lKrIWe3QaotbHZ0eJSC57UNf2mJTRAoh")
\# display in default browser
browseURL(baseball)}





}
\references{
\href{https://blog.usejournal.com/host-static-images-for-your-apps-or-website-on-google-drive-hotlink-to-gdrive-images-358d6fcf8ef7}{Blog post by P. Aboyi}
}
