% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-cleaning-funs.R
\name{col_clean}
\alias{col_clean}
\title{Remove empty columns}
\usage{
col_clean(df)
}
\arguments{
\item{df}{Tibble (or data frame)}
}
\description{
#' Returns a new data frame with \strong{columns} containing purely \code{NA} values removed
}
\examples{
# Create a table with two empty rows and one empty column
messy_table <- tibble::tibble(
  a = c(1:5, NA, 7:8, NA, 10),
  b = c(11:15, NA_character_, 17:18, NA_character_, 20),
  c = NA,
  d = c(sample(letters, 5), NA, "p", "j", NA, "Q")
)

# columns only; leave rows
col_clean(messy_table)

# rows only; leave columns
row_clean(messy_table)

# rows and columns
df_clean(messy_table)
}
