% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_clay_d_spacings.R
\docType{data}
\name{common_clay_d_spacings}
\alias{common_clay_d_spacings}
\title{A tibble of peaks often used in qualitative identification
of X-ray diffraction peaks for clay-bearing soils}
\format{
A tibble
}
\source{
\href{http://webmineral.com/}{WebMineral database}
}
\usage{
common_clay_d_spacings
}
\description{
A tibble of peaks often used in qualitative identification
of X-ray diffraction peaks for clay-bearing soils
}
\keyword{datasets}
