% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_excel_date.R
\name{convert_excel_date}
\alias{convert_excel_date}
\title{Convert an Excel date to \code{POSIXct} object}
\usage{
convert_excel_date(x)
}
\arguments{
\item{x}{Integer corresponding to Excel convention, where day 1 is Jan 1, 1900}
}
\value{
POSIXct object
}
\description{
Convert an Excel date to \code{POSIXct} object
}
\examples{
convert_excel_date(1) # First day of Excel time scale
convert_excel_date(46113) # April 2026
}
