% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-cleaning-funs.R
\name{df_clean}
\alias{df_clean}
\title{Remove columns or rows containing only NA values}
\usage{
df_clean(df, type = c("both", "rows", "cols"))
}
\arguments{
\item{df}{Tibble (or data frame)}

\item{type}{Character (length 1). Whether to remove empty rows, columns, or both.}
}
\value{
Tibble (or data frame) with any empty columns/rows removed
}
\description{
Remove columns or rows containing only NA values
}
