% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm_docx.R
\name{ecm_docx}
\alias{ecm_docx}
\title{Customized word document output}
\usage{
ecm_docx(...)
}
\arguments{
\item{...}{arguments passed to \code{rmarkdown::word_document()}}
}
\description{
Customized word document output
}
