% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm_timestamp.R
\name{ecm_timestamp}
\alias{ecm_timestamp}
\title{Print a time stamp including specific information}
\usage{
ecm_timestamp()
}
\value{
character vector with formatted weekday abbreviation, date, and time
}
\description{
Calls \code{\link[base:Sys.time]{base::Sys.time()}} with a particular \code{format} argument. Use for keeping track of .Rmd knitting.
}
\examples{
ecm_timestamp()
}
