% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecm_xaringan.R
\name{ecm_xaringan}
\alias{ecm_xaringan}
\title{Thin wrapper around \code{xaringan::moon_reader()}}
\usage{
ecm_xaringan(ratio = "16:9", seal = TRUE, ...)
}
\arguments{
\item{ratio}{Character. Aspect ratio of slides, specified as "width:height"}

\item{seal}{Logical. Auto-generate title slide?}

\item{...}{Other arguments passed to \code{xaringan::moon_reader()}}
}
\value{
Renders HTML slide show
}
\description{
A custom output format for HTML presentations
}
