% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_plot.R
\name{export_plot}
\alias{export_plot}
\title{Save a plot in a variety of formats with a single call}
\usage{
export_plot(x, dir = "figs", formats = c("svg", "png", "pdf"), rds = TRUE, ...)
}
\arguments{
\item{x}{ggplot object to save. The file name is automatically populated from the object's name.}

\item{dir}{directory in which to save the plots, defaults to 'figs' i.e. a directory named 'figs' located in the top level of current working directory}

\item{formats}{character. Image-based file formats to include. Defaults to 'svg', 'png', and 'pdf'.}

\item{rds}{Logical. Save in binary .rds format? (Default \code{TRUE})}

\item{...}{Currently not used. Eventually re-factor so other arguments may be passed to \code{\link[cowplot:save_plot]{cowplot::save_plot()}}}
}
\value{
Writes files to disk and prints a message if successful
}
\description{
Convenient for keeping plots at the ready for presentations, papers, social media, etc.
}
