% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_paper_figs.R
\name{find_paper_figs}
\alias{find_paper_figs}
\title{Find all figures inserted using \code{insert_fig()} to allow easy uploading for manuscript submission}
\usage{
find_paper_figs(file)
}
\arguments{
\item{file}{\code{.Rmd} file to search}
}
\value{
character vector of file stems
}
\description{
Looks in the \code{.Rmd} for any calls to this function and extracts the figure file names
}
