% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_anova.R
\name{format_anova}
\alias{format_anova}
\title{format an anova table output from \code{car::Anova()}}
\usage{
format_anova(x, caption)
}
\arguments{
\item{x}{ANOVA table output from \code{car::Anova()}}

\item{nm}{table caption}
}
\value{
silently writes slides to disk
}
\description{
Often make multiple tables at once by building a tibble with columns \code{x} and \code{caption}
through other data wrangling exercises
}
