% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsize.R
\name{fsize}
\alias{fsize}
\title{Print a file size in a human-readable format}
\usage{
fsize(path, digits = 0)
}
\arguments{
\item{path}{quoted path to file}

\item{digits}{number of decimal places to pring}
}
\value{
The file size, formatted with appropriate prefix
}
\description{
Print a file size in a human-readable format
}
\seealso{
\code{\link[=file.size]{file.size()}}, \link[gdata]{humanReadable}
}
