% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_ppt_save.R
\name{gg_ppt_save}
\alias{gg_ppt_save}
\title{Save a plot usingpowerpoint-compatible dimensions}
\usage{
gg_ppt_save(
  plot,
  filename,
  aspect_ratio = c("16x9-w-title", "two-content", "16x9-full-page"),
  size_multiplier = 0.8,
  dpi = 600
)
}
\arguments{
\item{plot}{ggplot object}

\item{filename}{Full path to save}

\item{aspect_ratio}{Height / width}

\item{size_multiplier}{Value < 1 = larger text}

\item{dpi}{For raster graphics; defaults to 600}
}
\value{
sliently writes file to disk
}
\description{
Save a plot usingpowerpoint-compatible dimensions
}
