% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsize.R
\name{image_filesize}
\alias{image_filesize}
\title{Print a file size in a human-readable format}
\usage{
image_filesize(path, digits = 0)
}
\arguments{
\item{path}{quoted path to file}

\item{digits}{number of decimal places to pring}
}
\value{
The file size, formatted with appropriate prefix
}
\description{
Print a file size in a human-readable format
}
\seealso{
\link[base]{file.size}, \link[gdata]{humanReadable}
}
