% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_fig.R
\name{insert_fig}
\alias{insert_fig}
\title{Helper for inserting external figures into an R Markdown document}
\usage{
insert_fig(
  stem,
  type = dplyr::if_else(knitr::is_latex_output(), "pdf", "png"),
  base = c("figures")
)
}
\arguments{
\item{stem}{Character. Basename of figure, without extension.}

\item{type}{Character. Format of the file. Defaults to \strong{knitr} and \strong{dplyr} logic.}

\item{base}{Character. Defaults to my usual structure of having a top-level folder for figures, abbreviated \verb{figures/}, but could also be 'figs/'}
}
\description{
Wraps \code{knitr::include_graphics()} to easily access my custom directory
structure for paper writing'
}
