% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpg2png.R
\name{jpg2png}
\alias{jpg2png}
\title{Convert a JPG image to png format}
\usage{
jpg2png(path)
}
\arguments{
\item{path}{Character. Path to one or more images to convert.}
}
\value{
Writes a copy of the JPG file as a png in the same directory.
}
\description{
A common task for photos captured with iPhone
}
