% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_with_date.R
\name{knit_with_date}
\alias{knit_with_date}
\title{Custom Knit function for RStudio, re: Yihui Xie}
\usage{
knit_with_date(input, ...)
}
\arguments{
\item{input}{the input file path}

\item{...}{other arguments passed to rmarkdown::render()}
}
\description{
Adds date the file was compiled to the file name. See https://bookdown.org/yihui/rmarkdown-cookbook/custom-knit.html
}
