% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ppt_table.R
\name{make_ppt_table}
\alias{make_ppt_table}
\title{Save a formatted ANOVA table to a powerpoint slide}
\usage{
make_ppt_table(x, template, output)
}
\arguments{
\item{x}{formatted data frame, typically returned by \code{format_anova()}}

\item{template}{path to .pptx file, possibly containing a custom theme}

\item{output}{path to write new slides (should end in '.pptx')}
}
\value{
silently writes file to disk
}
\description{
Save a formatted ANOVA table to a powerpoint slide
}
