% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ecm_xaringan.R
\name{new_ecm_xaringan}
\alias{new_ecm_xaringan}
\title{Create a new set of \strong{xaringan} slides with my own options and boilerplate}
\usage{
new_ecm_xaringan(file, open = FALSE)
}
\arguments{
\item{file}{path to save the new presentation}

\item{open}{whether to open the file in the R Studio IDE}
}
\value{
creates file and (if specified) opens in R Studio IDE
}
\description{
Create a new set of \strong{xaringan} slides with my own options and boilerplate
}
