% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_new_notes.R
\name{new_html_notes}
\alias{new_html_notes}
\title{Create a new Rmd document for note-taking purposes}
\usage{
new_html_notes(file, open = FALSE)
}
\arguments{
\item{file}{path to save the new presentation}

\item{open}{whether to open the file in the R Studio IDE}
}
\value{
creates file and (if specified) opens in R Studio IDE
}
\description{
Includes with my own boilperplate YAML and chunk options, and a
basic CSS file for formatting
}
