% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_check.R
\name{pkg_check}
\alias{pkg_check}
\title{Tell user to install missing packages}
\usage{
pkg_check(pkgs)
}
\arguments{
\item{pkgs}{Character vector of packages to look for}
}
\value{
error message listing the missing packages
}
\description{
A helper function to allow fewer packages to be listed in Imports section of DESCRIPTION file
}
