% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarterly_report.R
\name{quarterly_report}
\alias{quarterly_report}
\title{test function from Yihui}
\usage{
quarterly_report(toc = TRUE)
}
\arguments{
\item{toc}{whether to use a table of contens}
}
\value{

}
\description{
test function from Yihui
}
