% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.files.R
\name{read.files}
\alias{read.files}
\title{Read multiple files into R}
\usage{
read.files(.x = NULL, read_fn = NULL, enframe = TRUE, value = "file")
}
\arguments{
\item{.x}{a vector of file paths to read}

\item{read_fn}{un-quoted file reading function (specific to the file type; i.e. \code{read_csv} or \code{vcgImport})}

\item{enframe}{whether to create a tibble with a name column and a list-column of the files}

\item{value}{column name for the files held (ignored if \code{enframe = FALSE})}
}
\value{
a list or tibble of the files. Use \code{\link[=gather.files]{gather.files()}} to give them nice name before importing.
}
\description{
Wraps \code{purrr::map()} with an option to specify whether the list of files \code{.l}
should be compiled with \code{\link[tibble:enframe]{tibble::enframe()}}. Use after \code{\link[=gather.files]{gather.files()}} to
pipe in a vector of file paths with human-readable names.
}
