% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xrd.R
\name{read_xrd}
\alias{read_xrd}
\title{Parse a Panalytical XRDML file}
\usage{
read_xrd(path, scaled = TRUE, min_tth = 4, max_tth = 40)
}
\arguments{
\item{path}{Character. full path to file}

\item{scaled}{Logical. Whether to normalize intensity to maximum value (default \code{TRUE})}

\item{min_tth}{minimum value for 2-theta to keep (default 4)}

\item{max_tth}{maximum value for 2-theta to keep (default 40)}
}
\value{
Tibble with 2 columns
}
\description{
Produces a tibble of intensity and 2-theta vales
}
