% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-cleaning-funs.R
\name{row_clean}
\alias{row_clean}
\title{Remove empty rows}
\usage{
row_clean(df)
}
\arguments{
\item{df}{Tibble (or data frame)}
}
\value{
Tibble (or data frame)
}
\description{
Returns a new data frame with \strong{rows} containing purely \code{NA} values removed
}
