% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fig.R
\name{save_fig}
\alias{save_fig}
\title{helper for saving figures with GNU Make}
\usage{
save_fig(plot, pdf_path = NULL, ht = NULL, wdth = NULL, ...)
}
\arguments{
\item{plot}{a \strong{ggplot2} object}

\item{pdf_path}{path to which plot should be saved; if \code{NULL}, will call \code{commandArgs(trailingOnly = TRUE)[[1]]} to accept command line argument}

\item{ht}{height in inches, defaults to 9*0.4=3.6}

\item{wdth}{width in inches, defaults to 16*0.4=6.4}

\item{...}{other arguments passed to \code{ggsave}}
}
\value{
Silently writes files to disk and prints message
}
\description{
Export a \strong{ggplot2} object as a pdf and a png
}
\details{
Also saves in \code{./presentations/presentation-figs} for convenient use with \strong{xaringan} HTML slide shows
}
