% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_psa_options.R
\name{set_psa_options}
\alias{set_psa_options}
\title{Shortcut to set PSA equipment}
\usage{
set_psa_options()
}
\value{
Sets options silently
}
\description{
Sets all relevant global options for running a particle size analysis using ASI 468 glassware, tins, etc.
}
