% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-cleaning-funs.R
\name{strip_empty}
\alias{strip_empty}
\title{Remove columns or rows containing only NA values}
\usage{
strip_empty(df, type = c("both", "rows", "cols"))
}
\arguments{
\item{df}{Data frame.}

\item{type}{Character (length 1). Whether to remove empty rows, columns, or both.}
}
\value{
Data frame with any empty columns/rows removed
}
\description{
Remove columns or rows containing only NA values
}
