% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ecm_bw.R
\name{theme_ecm_bw}
\alias{theme_ecm_bw}
\title{Slight modifications to built-in themes}
\usage{
theme_ecm_bw()
}
\value{
extensible theme function
}
\description{
Based on \code{theme_bw()} from \strong{ggplot2}
}
\examples{
library(ggplot2)
library(ecmfuns)
ggplot(mtcars, aes(disp, mpg, color = cyl))+
 geom_point()+
 theme_ecm_bw()
}
