% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ecm_scatter.R
\name{theme_ecm_scatter}
\alias{theme_ecm_scatter}
\title{Personal preferences for a quick scatterplot}
\usage{
theme_ecm_scatter()
}
\value{
extensible theme function
}
\description{
Based off Claus Wilke's \code{\link[cowplot:theme_cowplot]{cowplot::theme_cowplot()}} and \href{https://wilkelab.org/ggtext/articles/introduction.html}{\strong{ggtext}} package
}
\examples{
library(ggplot2)
library(ecmfuns)
ggplot(mtcars, aes(disp, mpg, color = cyl))+
 geom_point()+
 theme_ecm_scatter()
}
