% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_abbs.R
\docType{data}
\name{time_abbs}
\alias{time_abbs}
\title{Format reference for dates and times}
\format{
Tibble with columns containing information about each element or
formatting string
}
\usage{
time_abbs
}
\description{
Easier than using R documentation for \code{strftime()}, etc.
}
\keyword{datasets}
