% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful_hex_codes.R
\docType{data}
\name{useful_hex_codes}
\alias{useful_hex_codes}
\title{Handy reference for often-used hexidecimal color codes}
\format{
A tibble
}
\usage{
useful_hex_codes
}
\description{
Commonly used for diagrams and plots
}
\keyword{datasets}
