% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_extract.R
\name{hex_extract}
\alias{hex_extract}
\title{\lifecycle{experimental}}
\usage{
hex_extract(image)
}
\arguments{
\item{image}{A cimg object which is an image file previously imported to the R session into R with \code{imager::load.image} (.png or .jpg)}
}
\value{
A character vector of length 1, containing average hexidecimal color
of all pixels in the image
}
\description{
Compute the average hex color of an image file
}
\details{
Use this function to quickly extract the average color of an image for display
on screens. Helpful for visualization purposes and making realistic drawings
in vector graphics software such as Adobe Illustrator.
}
