% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_notes.R
\name{html_notes}
\alias{html_notes}
\title{Rmd output format for note-taking. Renders to HTML output.}
\usage{
html_notes(
  css = NULL,
  toc = TRUE,
  number_sections = FALSE,
  highlight_style = "github",
  theme = NULL,
  ...
)
}
\arguments{
\item{...}{arguments passed on to \code{\link[rmarkdown:html_document]{rmarkdown::html_document()}}}
}
\value{
HTML output format function; passes arguments to Pandoc via \strong{knitr}
and \strong{rmarkdown}
}
\description{
Includes my own boilerplate YAML and chunk options, and a basic CSS file for
formatting
}
