% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/today_ecm.r
\name{today_ecm}
\alias{today_ecm}
\title{Return today's date in a professional format}
\usage{
today_ecm(month_abb = TRUE)
}
\arguments{
\item{month_abb}{Whether the month should be abbreviated or printed in full}
}
\value{
A professionally formatted character string containing the current system date.
}
\description{
The \strong{lubridate} package does an excellent job of parsing dates which are in messy formats back to POSIX format. However it is not so good for printing dates in the way the people typically read or think about them. This function prints the current date in a format more amenable to professional reports and presentations.
}
